breaking up state schemas into smaller more modular files
