+++ src/state-schemas/eval-frame-only-state-schemas.js
@@ -1,0 +1,102 @@
+import {
+  mirroredStateProperties,
+  mirroredCellProperties,
+} from '../state-schemas/mirrored-state-schema'
+
+import { newCellFromSchema } from '../state-schemas/state-prototype-from-schema'
+
+const DEFAULT_PANE_HEIGHT = Math.round(0.4 * document.documentElement.clientHeight)
+
+export const historySchema = {
+  type: 'object',
+  properties: {
+    cellId: { type: ['integer', 'null'] },
+    content: { type: 'string' },
+    historyId: { type: 'integer' },
+    historyType: {
+      type: 'string',
+      enum: [
+        'CELL_EVAL_VALUE',
+        'CELL_EVAL_EXTERNAL_RESOURCE',
+        'CELL_EVAL_INFO',
+        'APP_INFO',
+        'SAVED_REP',
+        'CONSOLE_EVAL',
+        'SNIPPET_EVAL',
+      ],
+    },
+    lastRan: { type: 'integer' },
+    value: {},
+  },
+  additionalProperties: false,
+}
+
+export const evalFrameOnlyCellProperties = {
+  asyncProcessCount: {
+    type: 'integer',
+    minimum: 0,
+    default: 0,
+  },
+  evalStatus: {
+    type: 'string',
+    enum: ['UNEVALUATED', 'PENDING', 'ASYNC_PENDING', 'SUCCESS', 'ERROR'],
+    default: 'UNEVALUATED',
+  },
+  hasSideEffect: {
+    type: 'boolean',
+    default: false,
+  },
+  rendered: {
+    type: 'boolean',
+    default: false,
+  },
+  value: {
+    // empty schema because `value` can be anything
+    default: undefined,
+  },
+}
+
+export const evalFrameCellSchema = {
+  type: 'object',
+  properties:
+    Object.assign({}, mirroredCellProperties, evalFrameOnlyCellProperties),
+  additionalProperties: false,
+}
+
+export const evalFrameOnlyStateProperties = {
+  cells: {
+    type: 'array',
+    items: evalFrameCellSchema,
+    default: [newCellFromSchema(evalFrameCellSchema, 0)],
+  },
+  externalDependencies: {
+    type: 'array',
+    default: [],
+  },
+  history: {
+    type: 'array',
+    items: historySchema,
+    default: [],
+  },
+  paneHeight: {
+    type: 'integer',
+    default: DEFAULT_PANE_HEIGHT,
+  },
+  userDefinedVarNames: {
+    type: 'array',
+    items: { type: 'string' },
+    default: [],
+  },
+  viewMode: {
+    type: 'string',
+    enum: ['EXPLORE_VIEW', 'REPORT_VIEW'],
+    default: 'REPORT_VIEW',
+  },
+}
+
+export const evalFrameStateSchema = {
+  type: 'object',
+  properties:
+    Object.assign({}, mirroredStateProperties, evalFrameOnlyStateProperties),
+  additionalProperties: false,
+}
